requirements.txt
Pillow==8.1.0
numpy==1.20.0
